+++ parler/cache.py
@@ -5,6 +5,7 @@ These functions are used internally by django-parler to fetch model data.
 Since all calls to the translation table are routed through our model descriptor fields,
 cache access and expiry is rather simple to implement.
 """
+import django
 from django.core.cache import cache
 from django.utils import six
 from parler import appsettings
@@ -13,11 +14,9 @@ from parler.utils import get_language_settings
 if six.PY3:
     long = int
 
-try:
-    # In Django 1.6, a timeout of 0 seconds is accepted as valid input,
-    # and a sentinel value is used to denote the default timeout. Use that.
-    from django.core.cache.backends.base import DEFAULT_TIMEOUT
-except ImportError:
+if django.VERSION >= (1,6):
+    DEFAULT_TIMEOUT = cache.default_timeout
+else:
     DEFAULT_TIMEOUT = 0
 
 
@@ -37,7 +36,7 @@ def get_object_cache_keys(instance):
     """
     Return the cache keys associated with an object.
     """
-    if not instance.pk or instance._state.adding:
+    if instance.pk is None or instance._state.adding:
         return []
 
     keys = []
@@ -57,7 +56,7 @@ def get_translation_cache_key(translated_model, master_id, language_code):
     """
     # Always cache the entire object, as this already produces
     # a lot of queries. Don't go for caching individual fields.
-    return 'parler.{0}.{1}.{2}.{3}'.format(translated_model._meta.app_label, translated_model.__name__, long(master_id), language_code)
+    return 'parler.{0}.{1}.{2}.{3}'.format(translated_model._meta.app_label, translated_model.__name__, master_id, language_code)
 
 
 def get_cached_translation(instance, language_code=None, related_name=None, use_fallback=False):
@@ -74,7 +73,12 @@ def get_cached_translation(instance, language_code=None, related_name=None, use_
     if not values:
         return None
 
-    translation = translated_model(**values)
+    try:
+        translation = translated_model(**values)
+    except TypeError:
+        # Some model field was removed, cache entry is no longer working.
+        return None
+
     translation._state.adding = False
     return translation
 
@@ -120,8 +124,14 @@ def _get_cached_values(instance, translated_model, language_code, use_fallback=F
         # Allow to return the fallback language instead.
         if use_fallback:
             lang_dict = get_language_settings(language_code)
-            if lang_dict['fallback'] != language_code:
-                return _get_cached_values(instance, translated_model, lang_dict['fallback'], use_fallback=False)
+            # iterate over list of fallback languages, which should be already
+            # in proper order
+            for fallback_lang in lang_dict['fallbacks']:
+                if fallback_lang != language_code:
+                    return _get_cached_values(
+                        instance, translated_model, fallback_lang,
+                        use_fallback=False
+                    )
         return None
 
     values['master'] = instance