+++ parler/tests/testapp/models.py
@@ -118,9 +118,11 @@ class ProxyModel(ProxyBase):
 class DoubleModel(TranslatableModel):
     shared = models.CharField(max_length=200, default='')
 
+
 class DoubleModelTranslations(TranslatedFieldsModel):
     master = models.ForeignKey(DoubleModel, related_name='base_translations')
     l1_title = models.CharField(max_length=200)
+
 
 class DoubleModelMoreTranslations(TranslatedFieldsModel):
     master = models.ForeignKey(DoubleModel, related_name='more_translations')
@@ -130,3 +132,19 @@ class DoubleModelMoreTranslations(TranslatedFieldsModel):
 class RegularModel(models.Model):
     # Normal model without translations. Test how replacing the field works.
     original_field = models.CharField(default="untranslated", max_length=255)
+
+
+class CharModel(TranslatableModel):
+    id = models.CharField(max_length=45, primary_key=True)
+
+
+class CharModelTranslation(TranslatedFieldsModel):
+    master = models.ForeignKey(CharModel)
+    tr_title = models.CharField(max_length=200)
+
+
+class ForeignKeyTranslationModel(TranslatableModel):
+    translations = TranslatedFields(
+        translated_foreign = models.ForeignKey('RegularModel'),
+    )
+    shared = models.CharField(max_length=200)