+++ parler/views.py
@@ -219,15 +219,15 @@ class LanguageChoiceMixin(object):
         """
         object = super(LanguageChoiceMixin, self).get_object(queryset)
         if isinstance(object, TranslatableModel):
-            object.set_current_language(self._language(object), initialize=True)
+            object.set_current_language(self.get_language(), initialize=True)
         return object
 
 
-    def _language(self, object=None):
+    def get_language(self):
         """
         Get the language parameter from the current request.
         """
-        return get_language_parameter(self.request, self.query_language_key, object=object, default=self.get_default_language(object=object))
+        return get_language_parameter(self.request, self.query_language_key, default=self.get_default_language(object=object))
 
 
     def get_default_language(self, object=None):
@@ -242,11 +242,12 @@ class LanguageChoiceMixin(object):
     def get_current_language(self):
         """
         Return the current language for the currently displayed object fields.
+        This reads ``self.object.get_current_language()`` and falls back to :func:`get_language`.
         """
         if self.object is not None:
             return self.object.get_current_language()
         else:
-            return self._language()
+            return self.get_language()
 
 
     def get_context_data(self, **kwargs):
@@ -268,10 +269,6 @@ class LanguageChoiceMixin(object):
         return get_language_tabs(self.request, current_language, available_languages)
 
 
-# Backwards compatibility
-TranslatableSingleObjectMixin = LanguageChoiceMixin
-
-
 class TranslatableModelFormMixin(LanguageChoiceMixin):
     """
     Mixin to add translation support to class based views.
@@ -290,7 +287,9 @@ class TranslatableModelFormMixin(LanguageChoiceMixin):
         Return a ``TranslatableModelForm`` by default if no form_class is set.
         """
         super_method = super(TranslatableModelFormMixin, self).get_form_class
-        if not (super_method.__func__ is ModelFormMixin.get_form_class.__func__):
+        # no "__func__" on the class level function in python 3
+        default_method = getattr(ModelFormMixin.get_form_class, '__func__', ModelFormMixin.get_form_class)
+        if not (super_method.__func__ is default_method):
             # Don't get in your way, if you've overwritten stuff.
             return super_method()
         else:
@@ -314,7 +313,9 @@ class TranslatableModelFormMixin(LanguageChoiceMixin):
 
 
     # Backwards compatibility
-    get_form_language = LanguageChoiceMixin.get_current_language
+    # Make sure overriding get_current_language() affects get_form_language() too.
+    def get_form_language(self):
+        return self.get_current_language()
 
 
 # For the lazy ones:
@@ -347,3 +348,7 @@ def _get_view_model(self):
     else:
         # Try to get a queryset and extract the model class from that
         return self.get_queryset().model
+
+
+# Backwards compatibility
+TranslatableSingleObjectMixin = LanguageChoiceMixin